Moved ControllerInterface, fixed bug where kerbal would be created w.o. description, started work on vessels
