+++ libs/KSPDB.b
@@ -1,0 +1,1 @@
+[1.0.0](000000000000000000000000000010)MISSIONS{}(000000000000000000000000000009)VESSELS{}(000000000000000000000000000241)KERBALS{30,,44696c646f,74727565,74727565,50696c6f74,4869726564,313a4b443a313a4b443a303a4b443a303a4b443a303a4b443a323032312d30322d30345430393a35333a30342e3737303838393830302b30313a3030,286e6f6e6529,286e6f6e6529,286e6f6e6529,302e30,66616c7365}(000000000000000000000000000403)CONCEPTS{31,,53504143454352414654,30,50656e69732050696e63686572,4e6f6e65,286e6f6e6529,454c4543545249433a56433a535352543a56433a494f4e5f504f57455245443a56433a56544f4c,4d554e5f4c494d495445443a56433a4b455242494e5f4c4f43414c3a56433a4b455242494e5f4c494d495445443a56433a4d554e5f4c4f43414c,323a4b443a333a4b443a303a4b443a303a4b443a303a4b443a323032312d30322d30345430393a35323a34382e3639373535333130302b30313a3030}(000000000000000000000000000009)CRASHED{}
