+++ src/kerbals/Condecoration.java
@@ -2,7 +2,7 @@ package kerbals;
 
 import missions.Mission;
 import other.*;
-import other.interfaces.ControllerInterface;
+import controller.ControllerInterface;
 import other.interfaces.KSPObjectDeletionEvent;
 import other.interfaces.KSPObjectListener;
 import other.util.Field;
